@@ -2,16 +2,20 @@ plugins {
     id 'com.android.application'
     id 'com.google.gms.google-services'
     id 'com.google.firebase.crashlytics'
+
+
 }
+
+
 
 android {
     namespace 'com.example.scanqrapp'
-    compileSdk 32
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.example.scanqrapp"
         minSdk 23
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -22,8 +26,19 @@ android {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+
+            //api key
+           // it.buildConfigField 'String', 'firebaseAPI', firebaseAPIuri
+
+
+        }
+        buildTypes.each {
+            it.buildConfigField 'String', 'firebaseAPI', firebaseAPIuri
         }
     }
+
+
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
@@ -36,38 +51,44 @@ android {
 
 dependencies {
 
-    implementation 'androidx.appcompat:appcompat:1.5.1'
-    implementation 'com.google.android.material:material:1.7.0'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.google.android.material:material:1.8.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.5.1'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 
+    // Navigation
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
+    implementation "androidx.fragment:fragment-ktx:1.5.5"
+
     //firebase
-    implementation 'com.google.firebase:firebase-abt:21.1.0'
+    implementation 'com.google.firebase:firebase-auth:21.1.0'
     implementation 'com.google.firebase:firebase-database:20.1.0'
+    implementation 'com.google.firebase:firebase-crashlytics:18.3.5'
     implementation 'com.google.firebase:firebase-storage:20.1.0'
-    implementation 'com.firebase:firebase-client-android:2.5.2'
+
 
     implementation 'com.android.support:multidex:1.0.3'
     implementation fileTree(dir: 'D:\\AppGallery\\appScan-\\app\\libs', include: ['*.aar', '*.jar'], exclude: [])
     implementation files('libs/poi-3.7.jar')
 
-    implementation 'androidx.core:core:1.8.0'
+ //   implementation 'androidx.core:core:1.9.0'
+
     implementation 'androidx.annotation:annotation:1.5.0'
-    implementation 'com.google.firebase:firebase-auth:21.1.0'
-
-    implementation 'com.google.firebase:firebase-crashlytics:18.3.5'
 
 
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.4'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
 
-    // Navigation
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
-    implementation "androidx.fragment:fragment-ktx:1.5.4"
+
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.9.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+
+    /*// Navigation
+     implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
+     implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
+     implementation "androidx.fragment:fragment-ktx:1.5.5"*/
 
     implementation 'com.github.yuriy-budiyev:code-scanner:2.3.2'
     implementation 'me.dm7.barcodescanner:zxing:1.9.13'
