@@ -1,22 +1,20 @@
 package com.example.scanqrapp;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.util.Patterns;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.ProgressBar;
-import android.widget.TextView;
 import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.AppCompatButton;
+import androidx.navigation.NavController;
+import androidx.navigation.fragment.NavHostFragment;
 
 import com.example.scanqrapp.databinding.ActivitySignUpBinding;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.FirebaseDatabase;
 
@@ -24,7 +22,7 @@ import java.util.Objects;
 
 public class SignUpActivity extends AppCompatActivity implements View.OnClickListener {
 
-    private TextView big_header;
+
     private AppCompatButton user_regBtn;
     private EditText editTextfullName;
     private EditText editEmail;
@@ -41,14 +39,20 @@ public class SignUpActivity extends AppCompatActivity implements View.OnClickLis
                 .inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
+
+        //add nav graph/ nav host to default view
+        NavHostFragment navHostFragment =
+                (NavHostFragment) getSupportFragmentManager().findFragmentById(R.id.nav_graph);
+        if (navHostFragment != null) {
+            NavController navController = navHostFragment.getNavController();
+        }
+
         mAuth = FirebaseAuth.getInstance();
 
         //implement user reg btn
         user_regBtn = (AppCompatButton) findViewById(R.id.user_regBtn);
         user_regBtn.setOnClickListener(this);
 
-        /*big_header = (TextView) findViewById(R.id.big_header);
-        big_header.setOnClickListener(this);*/
 
         editTextfullName = (EditText) findViewById(R.id.input_name);
         editEmail = (EditText) findViewById(R.id.reg_email);
@@ -56,23 +60,30 @@ public class SignUpActivity extends AppCompatActivity implements View.OnClickLis
 
         progressBar = (ProgressBar) findViewById(R.id.indeterminateBar2);
 
-        /*Navigation.findNavController(binding.getRoot())
-                .popBackStack();
-        Navigation.findNavController(R.id.iv_back).popBackStack()*/
+        binding.viewBack.setOnClickListener(view -> startActivity(new Intent(SignUpActivity.this, LoginActivity.class)));
     }
 
 
    @Override
     public void onClick(View view) {
-        /*case R.id.big_header:
-                //navigate to login page
-                startActivity(new Intent(this, MainActivity.class));*/
-       if (view.getId() == R.id.user_regBtn) {// get register activity
-           registerUser();
-       } else {
-           // Toast if user is unable to register account
-           Toast.makeText(SignUpActivity.this,"Try again Later",Toast.LENGTH_SHORT).show();
-       }
+       try{
+            if (view.getId() == R.id.user_regBtn) {
+                registerUser();
+                startActivity(new Intent(this, LoginActivity.class));
+            }else {
+                // Toast if user is unable to register account
+                progressBar.setVisibility(View.GONE);
+                Toast.makeText(SignUpActivity.this,"Try again Later",Toast.LENGTH_SHORT).show();
+            }
+
+        }catch (Exception e){
+           e.printStackTrace();
+
+           Log.d("debug", "onClick: ");
+           // startActivity(new Intent(SignUpActivity.this, SignUpActivity.class));
+
+        }
+
     }
 
     private void registerUser() {
@@ -118,45 +129,49 @@ public class SignUpActivity extends AppCompatActivity implements View.OnClickLis
             editEmail.requestFocus();
             return;
         }
+        try {
+            progressBar.setVisibility(View.GONE);
+            mAuth.createUserWithEmailAndPassword(reg_email, reg_pass)
+                    .addOnCompleteListener(task -> {
 
-        progressBar.setVisibility(View.GONE);
-        mAuth.createUserWithEmailAndPassword(reg_email,reg_pass)
-                .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()){
 
-                            User user = new User(input_name,  reg_email);
+                        if (task.isSuccessful()) {
 
-                            
+                            User user = new User(input_name, reg_email);
+
+                            FirebaseDatabase.getInstance(BuildConfig.firebaseAPI)
+                           
                                     .getReference("Users")
                                     .child(Objects.requireNonNull(FirebaseAuth.getInstance().getUid()))
                                     .setValue(user)
-                                    .addOnCompleteListener(new OnCompleteListener<Void>() {
-                                        @Override
-                                        public void onComplete(@NonNull Task<Void> task) {
+                                    .addOnCompleteListener(task1 -> {
 
-                                            if(task.isSuccessful()){
-                                                Toast.makeText(SignUpActivity.this, "user has been registered! ", Toast.LENGTH_LONG).show();
-                                                progressBar.setVisibility(View.VISIBLE);
-                                                Log.d("Error", Objects.requireNonNull(task.getException()).toString());
+                                        if (task1.isSuccessful()) {
+                                            Toast.makeText(SignUpActivity.this, "user has been registered! ", Toast.LENGTH_LONG).show();
+                                            progressBar.setVisibility(View.VISIBLE);
+                                            Log.d("Error", Objects.requireNonNull(task1.getException()).toString());
 
-                                            } else {
-                                                Toast.makeText(SignUpActivity.this, "registration failed, check your info and try again! ", Toast.LENGTH_LONG).show();
-                                                progressBar.setVisibility(View.GONE);
-
-                                            }
+                                        } else {
+                                            Toast.makeText(SignUpActivity.this, "registration failed, check your info and try again! ", Toast.LENGTH_LONG).show();
+                                            startActivity(new Intent(this, LoginActivity.class));
+                                            progressBar.setVisibility(View.GONE);
 
                                         }
+
                                     });
 
-                        } else{
+                        } else {
                             Toast.makeText(SignUpActivity.this, "failed to register", Toast.LENGTH_LONG).show();
                             progressBar.setVisibility(View.GONE);
                             Log.d("Error", Objects.requireNonNull(task.getException()).toString());
 
                         }
-                    }
-                });
+
+                    });
+        }catch (IllegalStateException e){
+            Toast.makeText(SignUpActivity.this,"Please Wait", Toast.LENGTH_LONG).show();
+            startActivity(new Intent(this, LoginActivity.class));
+
+        }
     }
 }
